:unicorn:Address compiler-specific errors and warnings
- Avoid a signed-unsigned comparison in `STPWorldPipeline`.
- Ensure enum type for shader memory type in `STPExtendedScenePipeline` provided is valid.
- Add extra parentheses in `STPScenePipeline` for clarity.
- Add dependent template specifier in `STPNullablePrimitive`.
- Remove `TypeCount` from `STPTextureType` because it is no longer useful.

+++ SuperTerrain+/Core/World/STPWorldPipeline.cpp
@@ -556,7 +556,7 @@ public:
 	*/
 	inline void mainWaitsWorkers(cudaStream_t main) const {
 		//record the status of all worker streams.
-		for (int worker = 0; worker < this->MemoryWorker.size(); worker++) {
+		for (size_t worker = 0u; worker < this->MemoryWorker.size(); worker++) {
 			STP_CHECK_CUDA(cudaEventRecord(this->MemoryWorkerSync[worker].get(), this->MemoryWorker[worker].get()));
 		}
 		//main stream should not continue until all workers are done
